Move BaseHTTPEndpoint to common.http.base_endpoint
