+++ starlette_web/common/http/requests.py
@@ -1,0 +1,7 @@
+from sqlalchemy.ext.asyncio import AsyncSession
+from starlette.requests import Request
+
+
+class PRequest(Request):
+    user_session_id: str
+    db_session: AsyncSession
