+++ starlette_web/core/database.py
@@ -1,3 +1,4 @@
 from sqlalchemy.orm import declarative_base
 
+
 ModelBase = declarative_base()
